Refactor Hydra plugin
Move FileMeshAdapter class to its own files.
Remove CubePerMeshPoint plugin, which was only there for testing.
