+++ src/procedural/plugInfo.json
@@ -25,13 +25,6 @@
                         "primary": true,
                         "target": "usd"
                     },
-                    "CubePerMeshPointProceduralPlugin": {
-                        "bases": [
-                            "HdGpGenerativeProceduralPlugin"
-                        ], 
-                        "displayName": "CubePerMeshPoint",
-                        "priority": 0
-                    },
                     "GolaemProceduralPlugin": {
                         "bases": [
                             "HdGpGenerativeProceduralPlugin"
